+++ app/src/main/java/org/jboss/hal/client/runtime/host/HostColumn.java
@@ -67,6 +67,7 @@ import com.gwtplatform.mvp.shared.proxy.PlaceRequest;
 
 import elemental2.promise.Promise;
 
+import static elemental2.dom.DomGlobal.console;
 import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.toList;
 import static org.jboss.hal.client.runtime.configurationchanges.ConfigurationChangesPresenter.CONFIGURATION_CHANGES_ADDRESS;
@@ -194,7 +195,7 @@ public class HostColumn extends FinderColumn<Host> implements HostActionHandler,
                             actions.add(ItemAction.separator());
                         }
                         // TODO Add additional operations like :reload(admin-mode=true), :clean-obsolete-content or
-                        // :take-snapshot
+                        //  :take-snapshot
                         actions.add(new ItemAction.Builder<Host>()
                                 .title(resources.constants().reload())
                                 .handler(hostActions::reload)
@@ -232,6 +233,7 @@ public class HostColumn extends FinderColumn<Host> implements HostActionHandler,
             Host host = event.getHost();
             ItemMonitor.stopProgress(Ids.host(host.getAddressName()));
             event.getServers().forEach(server -> ItemMonitor.stopProgress(server.getId()));
+            console.log("### Refresh host column");
             refresh(RESTORE_SELECTION);
         }
     }